add active players command
